[bugfix] sizes are not passed correctly in scatter
Due to a limitation of the DUNE grid interface, sizes are not passed correctly
to the data handle in scatter, if data is sent on multiple entities. We won't
fix this issue, but circumvent it be declaring the ability to communicate only
for codimension 0.

Note: The grid can still communiate on vertices and edges if you do not require
      the size information in scatter. But this is no longer covered by the
      officially declared feature set.